+++ scratch2/draw.js
