+++ syntax/commands.js
